[FEATURE] Dispatch PSR-14 `ModifyTcaSelectFieldItemsEvent` in `itemsProcFunc`
`EXT:academic_jobs` provides following `itemsProcFunc` handler:

* `EmploymentTypeItems`
* `TypeItems`

which defines default items for fields (TCA or flexfom) using
these `itemsProcFunc` handlers. These are used in the backend
within the `FormEngine` and additionally dispatched manually
in the `JobController` to retrieve the available select items.

Previously, it was required to add them either directly not
having an influence to the ordering OR to implement a custom
`itemsProcFunc` replacing the shipped version and both ways
are not really nice or testable.

To provide an easier way for projects to modify the available
select items, the handler now dispatches the newly introduced
shared event `ModifyTcaSelectFieldItemsEvent` added to the
`EXT:academic_base` extension to reuse the same event in the
other extension later and having an easy way to modify these
thins within one event listener and a central event avoiding
the need to deal with douzen of dedicatly named events for
the same purpose.

Further, the event and handler defines proper phpstan array
shapes helping directly to get the structure of the passed
argument.

Dispatching the event and that the event can be used to
manipulate the items are covered with functional tests.