[TASK] ACE-105: Add `TableConfiguration to `EXT:academic_base`
This change adds the new `TableConfiguration` wrapper class
to `EXT:academic_base` providing access to reduces subset
of TCA information of a table required for other academic
extensions in object oriented way.

TYPO3 v13 provides the new `TcaSchema` not available in v12
and the reason why we introduces a small implementation on
our own to make access to TCA more bulletprove and reduce
recuring nasty array handling.

Some tests are added to cover the implementation.

Note usages to this class will be added in following changes.