[TASK] Introduce modern environment state manager
TYPO3 has different context environments, which comes
with different state and globals. Some services relates
to specific context, for example a frontend request
environment with a (global) request object and the
attributes along with the global TSFE and needs to
switch them for different configuration, for example
CLI commands or scheduler tasks.

The required work to build such environments is quite
some work and differs between TYPO3 versions and it's
important to have a deep understanding to do that right.

This change introduces a couple of services including
TYPO3 version based specialists to provide the ability
to provide a environment (EnvironmentBuilder), while
having a satisfing state mangaer to backup and restore
any environment in a stack to return to the orignal
state after a simulated environment was necessary to
build.

This allowes to create enforced FE environments with
always loaded TypoScript configuration required for
example for extbase based extensions or similar. The
implementation is splitted to avoid PHPStan issues
and baseline entries while having a concrete working
implementation for each TYPO3 version under the hood.

Functional tests has been choosen to cover the working
state in a automatic way, special for the dual core
support and to avoid classhes between versions.

The change itself is not used and acts as a preparation
step to replace places in other academic extensions to
have a working environment state management in place
and switch environments on the fly. Services are either
injectable or instancible using `GU::makeInstance()`,
returning the correct specialized services under the
hood.

Two internal PSR-14 events are added to allow custom
state backup and apply using the the `additionalData`
array of the `StateInterface`.

`EXT:academic_persons/Clases/Command/CreatesProfilesCommand`
is modified to backup and restore the environment state to
mitigate issues with pullated state. This is done one-time
globally and than for each iterated frontend user.